Threaded version
Found that discord will disconnect the bot if it is blocked for around a
minute which makes it unable to handle running a long batch script in a
blocking subprocess call. This version runs the script on a separate
thread. One issue introduced is the console no longer outputs the output
of the RunUAT script - this may not be necessary anyway, quick fix was
to output the contents of the call in batch to an output file in the log
directory.